mailkarantempEmailNator
The website clearly states: "Never use temporary mail for important information. It is a public email and can be accessed by anyone and your mail address is only temporary."

If you're interested in using this service, you can simply visit the website and click the "Generate New" button to create a temporary email address.